[fidlc] Fix out-of-order constant size resolution
This fixes an issue that is part of the "out-of-order" class of issues
in fidlc.

Previously, sized types (i.e. arrays, vectors, and strings) did size
resolution during the consume pass of compilation, which is okay if
the size constant has previously been consumed (i.e. constant is
defined earlier in the file or is located in an already-consumed
file). However, if a non-literal constant size value is defined in a
file that has yet to be consumed, the compiler will incorrectly fail on
being unable to parse the type size. This change moves constant size
resolution to the compilation pass to avoid this error.

Additionally, this change also addresses the inability to resolve
chained constant value definitions, i.e.:

    const uint32 DEPENDENT_CONST = ANOTHER_CONST;
    const uint32 ANOTHER_CONST = 5;

FIDL-298 #done

As side effects, this change also:
    (1) Merges constant type-checking with constant resolution into the
        same resolve pass,
    (2) Performs more granular type-checking on constants
    (3) Performs numeric limit checks on arithmetic constant types,
    (4) Makes significant progress on FIDL-304 as enum member value
        constants can now be properly resolved with ResolveConstant(),

Test: Manual testing with handwritten examples for size type
resolution success/failure and chained definitions. Additionally, new
test cases have been included in this change to specifically test such
cases. Run fidl-compiler-tests to test both existing and new fixed
functionality.

Change-Id: I2ced046f103a9e0404bf33272aa3a3220263110a