[docs][symbolize] Make symbolization easier
This change makes some simple quality of life improvements to
symbolize-filter. Additionally it adds a simple doc outlining
how to symbolize anything in zircon. These docs are valid for
both ASan and non-ASan crashes now but these docs outline the
*only* way to symbolize ASan crashes. This is intended to
be the new default symbolizer in zircon and a PSA will precede
submitting this change telling people to switch to the new
symbolizer.

Bug: TC-21 #comment zircon changes for transition done.
Test: By hand
Change-Id: I990499b83b8c31dcfd683c5f4dca38131ce26ebb