[media][tests] Renderer gain callbacks
The GainControl interface bound from AudioRenderer is currently
incomplete -- the OnGainMuteChanged event is not yet supported.
In order to test upcoming changes to AudioRenderer::SetMute, this
CL adds event support for the GainControl binding -- specifically
bindings created from an AudioRenderer interface. Naturally, this
CL includes a significant number of new tests to validate the added
feature.

Test: Build, CQ, and a new GainControlTest set in audio_core_tests.

Change-Id: Icfa94d0767170bf74d050da3d9b901b2acdbd10d