[cobalt-client]: Revamp to reduce allocations.
Metric information is known at compile time, for this reason
we are moving that information into compile data to prevent
overloading the start up. Plus has the benefit that allows
everything to be allocated contigously on the stack, which
reduces the number of indirections.(vs using a vector).

Instead of the Collector owning the memory, the actual
metrics (Histogram, Counter) will own their memory, and
will allow the Collector to register with them.

This means the usage will become:
MyHistogram<BucketCount> hist =
MyHistogram<BucketCount>(HistogramOptions, &collector);

If no collector is supplied, then the metric is simply a
getter and setter for counters and histograms. Though,
registering local and remote metrics with the Collector
allows it to provide Flushing mechanisms to cobalt,
and expose local values as a PseudoFile(s) + *PseudoDir.

All the Collector will have is a collection of
FlushInterface, which allows it to control the Flushing lifecycle,
and the only vector that exists is for the number of FlushInterface.

TEST=cobalt-client-test

Change-Id: I47820dc72f520b22263e63ec84cb48c9f37987ed