[appmgr] implement fuchsia.process.Resolver
Implementing the process resolver service provides a step toward being able
to launch command line binaries from packages other than /system. This is not
on the path to tools as components - that will be a larger effort - but this
path intends to unblock /system deprecation and enable on-demand installation
of developer tools.

Bug: PKG-54 #comment implement fuchsia.process.Resolver in appmgr
Test: echo '#!resolve fuchsia-pkg://fuchsia.com/system_image#bin/runtests' > /tmp/runtests; /tmp/runtests
Change-Id: Iebb0f3d599bd3a18f1a679d80e3d1439ffce500b