[rust] Define object properties and type safe accessors
Adds definitions of the object properties to fuchsia-zircon-sys and defines type safe
wrappers in fuchsia-zircon. The wrappers for getting and setting properties rely on
the implementation of unsafe traits to indicate what the correct type is for a property,
and whether it is permitted to call get or set on the specific property.

The |unsafe_handle_properties| macro is defined to avoid excessive boilerplate when writing
per object wrappers for properties.

Test: run-test fuchsia_zircon_lib_test_rustc

Change-Id: I93e006709a65bdd6d16c563902a62607c649e19d