[set] Add singular --board and --product flags, deprecated plurals
The plural forms of --boards and --products are conceptually odd - the
union of two boards or products will generally speaking not be a valid
configuration. This adds the singular forms of these flags which are
likely most people want most of the time and documents the plural forms
as deprecated. Both forms still work for now.

Test:

BLD-318 #comment add singular flags

Change-Id: Ie1902386e2cbe4010d8e575881ef53450a2259d7