[third_party][lz4] add fuzzers for lz4 library.
This CL adds basic fuzzers for the lz4 library.

There are two fuzzers being added. Once is a "roundtrip" fuzzer, that
will try to compress the input data, decompress the result, and check
that they are equal. The other fuzzer just tries to decompress random
data and crashes right away (until we fix it).

TEST=`fx fuzz start lz4.{fuzzname}` in a pixelbook running fuchsia.

Change-Id: Ic3b9ab98e3c8cb1d345605ad3b0b667dac8b8c5f