[debugger] Revamped setting get/set
Added job setting schema with the dummy Job filters option.

Getting the values will now get the correct settings out of the given
command context.

Setting will now also check for jobs schemas. It works for different
jobs contexts, but the interface is awkward, as you need to explicitly
qualify in order to be able to set a setting.

TEST=unit, manual

Change-Id: I854ebaf4787d0e41f61817ed30e7ab9770e786f3