[callback] Document Waiter class and refactor as a state machine
Minor change of behavior: Accumulator::Update is not called anymore when
its result is guaranteed to be ignored (eg. when the waiter has been
cancelled). This is only observable if an accumulator has external
side-effects in its Update method, which is not the case of any of the
existing accumulators.

Test: fx shell "runtests /pkgfs/packages/garnet_lib_tests/*/test -- --gtest_filter=*Waiter*"
Bug: LE-382 #done
Change-Id: Ic36886adbca3f97e1aacabbf8158404080340da3