[debugger] Avoid reallocating frames vector.
Previously Thread.GetFrames would return a copy of the frames vector
which is constructed from the backing vector that owns the pointers.

As we add more complicated thread stepping code, we're adding more calls
to GetFrames() so requiring a copy every time feels wasteful (though is
surely not measurable). This adds a cached copy so we can return a const
reference from GetFrames().

TEST=unit

Change-Id: Iff3aca8f03a5b4e692cb924f6d99b5a40fcae90f