[kernel][thread] Strongly type user_thread member of thread_t
All threads in zircon (user mode as well as kernel) are based on the
thread_t structure.  User mode threads have an additional
ThreadDispatcher class layer which encapsulates a thread_t instance.
The thread_t structure maintains a back pointer (the field is named
"user_thread") to this ThreadDispacher instance for user mode threads,
or nullptr in the case that the thread is a kernel thread.

In the past, the type of the user_thread member of the thread_t
structure has been a void*.  This was required to maintain
compatibility with C code which needed to deal with thread_t
instances.  Now that all of the code which needs to understand
thread_t instances is C++, we can go ahead and change the void* to be
a strongly typed ThreadDispatcher* instead.

In addition, there was a place in the kernel unit tests where a
thread_t* was being stuffed into the user_thread member instead of a
ThreadDispatcher* which was caught by the conversion to a strong type.
Tweak the test and the "user callback" signature to keep this clean
and not mix incompatible type pointers.

Tests: Build, user-mode unit tests, and kernel-mode unit tests.
Change-Id: I52cdf40ac9ea5a1f5122499dc078d6b3dbfb4fe9