[futex] Soft transition to updated futex API: Step 2
Perform a soft transition to an updated futex syscall API which
adds support for building sync objects which implement priority
inherance out of futexes.

The steps involved in this transition are as follows.
1) Introduce new versions of all syscalls whose signatures are
   changing with a _deprecated suffix.  These versions are simple
   passthrus to the old functionality.
2) Update all non-zircon users to use the _deprecated versions.
3) Update the syscall signatures in zircon, and update all zircon
   users.  Retain the deprecated flavors for now.
4) Update all non-zircon users to use new versions of the syscalls.
5) Remove the _deprecated versions of the syscalls from zircon.

Step #1 was completed with I61ce7ec228cf0da847d8f1065b368edd624bd195
This is part of step #2.

Tests: build garnet, runtests on NUC
Change-Id: I069e81a3f46e08c2fe2b0d298acf472de5031f7e