[mexec] Create a syscall to patch zbi.
Currently the mexec call is responsible for both patching the
bootdata with system information and for booting into the next
next system.

This patch introduces the zx_system_mexec_payload_get syscall which
returns an incomplete zbi containing a payload that should be appended
to the zircon.zbi produced by the build to successfully mexec the
zircon.zbi image.

Aside from simplifying the machinery of the zx_system_mexec syscall and
making that syscall more generic, this division of functionality
precludes the kernel from manipulating VMOs passed in from userspace.
I.e. userspace is now responsible for constructing a full bootable
image.

NOTE: For now, this syscall is unused and the system maintains its
current functionality. In the future, we'd like to transition
mexec to using this syscall.

Test: Compiled.
Change-Id: If18dd8d90fb956c7a055fddf92a48f962f40638a