[debugger] Changed set logic to be more intuitive.
Now it can find settings at the appropiate level without having to
explicitly qualify. This will pose a problem with ambiguous setting
naming, but that can be addressed in another CL.

You can still qualify, and can be either a NOP (if the setting only
lives in a specific context, say, job) or can set it at the highest
level possible.

TEST=manual

Change-Id: Ie3dc23b52515c70c54b3efd696f48082d21c740b