[async-promise] Integrate fit::promise and async_dispatcher_t.
Introduces async::Executor, an implementation of fit::executor
backed by an async_dispatcher_t.  This allows fit::promises to
run on our message loops.

This library should ideally be merged with libasync-cpp in Zircon
but we don't yet have use of libc++ there so I am landing it in
Garnet for now to unblock consumers there but will move it down
to Zircon once that is possible.  Consequently, this library
follows Zircon code style conventions.

Test: async_promise_unittests
Change-Id: Idf3544997d3d1292dd105e476631898a4c0c3f52