[cobalt-client] Remove EventBuffer.
Remove EventBuffer<T> because after changing to the new cobalt API,
and moving towards Remote/Local data, it was just a wrapper for
a type T. It did provide the exclusive TryFlush/CompleteFlush
functionality BUT this is not necessary.

Because no of the Flush operations at a Collector level may overlap,
the Collector already provides a barrier to prevent concurrent Flushing,

For the same reason the FlushInterface now only provides:
 - Flush
 - UndoFlush

And Flush returns True if it suceeds, and False if it fails, since the
synchronization is handled at an upper level. Which means the
FlushResult enum was also removed.

Benefits:
- Less Code
- Less Abstractions

Also removed some public internal aliases, RemoteCounter::Type.

TEST=cobalt-client-test

Change-Id: Id840bc0f7e05c6d78941f4e5e63f834a8b089146