[debugger] Generalize Location symbol storage.
Previously the Location object had a function() member that's the symbol
for the function the code location is in. This patch generalizes this to
be a symbol. This will support future indexing of global variables,
though there is no current change to functionality.

The function() member of Location is renamed symbol(). Callers are
updated.

Some additional error messaging and comments are added relating to TLS
values. TLS is currently not supported.

The ModuleSymbolsImpl can now create non-function locations given
Variable matches provided by the ModuleSymbolIndex. This is not
currently running however as no non-function symbols are indexed (this
will be done in a folowup).

Also done in a followup will be looking up these symbols to resolve
variables. This phase will also include adding tests for the lookup of
non-Function symbols.

TEST=unit

Change-Id: Ib70256beb7a3cc7e38295e8c1c3482e773fd0389