[debugger] Added debug-info flag to zxdb
This flag enables the debugger to know that it can output further debug
information.

The first application of this is to make the MessageLoop output which
tasks is executing, which was very opaque before. This required to
change the callsites of every PostTask call.

Added a FileLineFunction class that closely mirrors symbol's FileLine. I
started moving that one to debug_ipc, but there were to many changes and
it didn't seem worth while.

TEST=manual

Change-Id: I006c67995210944989f63bdb21ac6f0616463bbd