[set] Allow adding packages to available / preinstall / monolith sets
This adds flags to 'fx set' to configure additional packages to appear
in the available, preinstall, and monolith sets.  These are additive on
top of the packages specified by the board & product configuration.

This also marks the 'products' flag deprecated.

A subtle side effect of the way these flags are defined is that setting
these sets does not suppress the product inference.  This means that
once everything migrates to these flags and away from packages, there
will always be a product defined by 'fx set' similar to how there is
currently always a board specified.

Test: fx set with these args, examine args.gn

BLD-240 #comment add ways to add to available/preinstall/monolith sets

Change-Id: I1204a0388de30118cd2cfd471e770914ea5c1ab0