[zx_thread_read/write_state] Removed arch thread_t state checks.
The arch specific getter/setter of some registers (vector/debug) where
making checks to see if the thread_t (LK thread state) wasn't running.
This is not necessary, as there are previous checks that verifies that
the ThreadDispatcher state is SUSPENDED. There is a contract that when
the suspended callback is issued, it is safe to read the register state,
so this should work.

As a TODO, the ARM debug register state getter queries the HW each time,
which is not ideal. Normally you'd track that information and keep up
the date. That will come in a future CL.

TEST=Unit test passes
ZX-2508 # Bug that tracks the complete race condition for registers.

Change-Id: I37396460df66530814eecdfaf09aa313eff408a9