[upci] Simplify the pciroot structure for ecam
This removes the inner ecam structure that used to be part of
pci_platform_info_t and moves end_bus_num to top level. Since
ecam base was only used for debugging it isn't carried over into
the bus driver anymore. This means the bus driver can now determine
if it has an ecam available by simply checking if a valid handle was
passed to it by the pciroot protocol upon calling GetPciPlatformInfo.

Test: Confirmed config output with upci, and checked booting fuchsia
with a gui for kpci.

Change-Id: I24760bf3faed4a468c910b147297b20f60adf9fa