[garnet] wlan-ap-smoke-test
The test looks for MAC roles of each interface. On all interfaces that
support the AP role, it goes one at a time to startup an AP.
If there are other interfaces that support client role,
it has them scan, connect and disconnect to this AP, one at a time.
When complete it stops the AP. Test results are collected in each
step of the process and printed in the end.

Bug: NET-1731

Test: fx run-test wlan_ap_smoke_test_bin_test_rustc TestNetwork 6
Test: fx run-test connectivity_testing_lib_test_rustc

Change-Id: I8dbe43e183009fd32fc693ed049d9260557a64ec