[guestmgr] OnShutdown notification in vsock interface
The HostVsockEndpoint needs a mechanism to know when a connection has terminated
in order to reuse ephemeral ports. Previously it would keep a copy of one side
of the socket and listen for the host client closing the connection, but this
would prevent a host client from learning about shutdowns initiated from a guest.

Using an event on the HostVsockEndpoint fidl enterface creates an additional
channel of shutdown information allowing the manager to receive the OnShutdown
event, and the client to still receive PEER_CLOSED on its socket/channel.

TEST: run-test guestmgr_unittests,guest_integration_tests
Change-Id: I64e0eff9bdf8282f85283e2f69326c52b021f94b