[debugger] Index global variables.
Adds globals, class statics, and file statics to the symbol index.
Previously the index only contained functions.

These are not currently consumed, but will allow evaluating global
variables in a future patch.

This updates the checked-in symbols file so that it can include the
tew types of symbols.

Does some slight cleanup of parent setting in DwarfSymbolFactory. The
parent DIE was being set redundantly in some code paths (the main
DecodeSymbol() function will set it if the type-specific ones don't set
it. The way function DIEs work with by merging the spec and
implementation is clarified a bit in the comments and this pattern is
duplicated for variables.

TEST=unit

Change-Id: Ic767a2a7323e4d0475209dd9e7dcb0dc71e11533