[media] Add Mute to wav_recorder
This (and Usage) should probably be the last features added
to wav_recorder before it does what signal_generator did:
split into two entities with wholly different intentions.
* The first is a return to its original, vastly-simpler form:
an EXAMPLE along the lines of simple_sine, but for audio input.
* The second is a TOOL along the lines of signal_generator,
but for audio input -- with a more complex collection of
assorted features such as capturer Gain/Mute/Ramping (even
on-the-fly) as well as Usage-related changes.

This CL adds to wav_recorder the ability to explicitly set or
clear the Capturer's Mute state. This is used when developing
or debugging features related to capture gain/mute/ramping.

wav_recorder is a media example that (in the absence of an input-
oriented tool) is used by developers for diagnostic purposes.

Test: build, CQ, manual validation

Change-Id: If17a0046e927a161d1607446840183650e830a40