[bootsvc] Introduce bootsvc program
The plan is for bootsvc to become the program launched by userboot
instead of devmgr.  bootsvc will stick around past early-boot and
provide bootfs and a loader that only loads from bootfs.  It will
launch a single program, which for now will be devmgr, but in the
future will likely be a component manager.

This change does not switch bootsvc to be userboot's next stage yet,
since some changes to devmgr are necessary.

ZX-3027 #comment Core of implementation

Test: Ran scripts/run-zircon-x64  -V -b -k  -q \
    $(pwd)/../buildtools/linux-x64/qemu/bin -c userboot=bin/bootsvc \
    -c bootsvc.next=bin/bootsvc-tests

Change-Id: I28c0810798610da77fc9c9252cc81f8f3f0be214