[kernel] VcpuDispatcher::Interrupt -> VcpuDispatcher::PhysicalInterrupt
Simplify the VcpuDispatcher API since there is only one caller of
Interrupt() and it always uses hypervisor::InterruptType::PHYSICAL.

This avoids a circularity in the #include dependencies between
modules, which is a problem for the new build system.  Perhaps in
future re-examine the layering so that kernel/lib/hypervisor does
not depend on kernel/object, which seems like a layering inversion.

Test: CQ
Change-Id: Ifd6723977fce257190cd75b85d1409f734f53d2a