[debugger] Global variable printing infrastructure.
This has no functionality change, it is infrastructure to support future
additions only.

Does some reorganization to the SymbolEvalContext so there is a place to
resolve global symbols, and the context necessary to do so (the
ProcessSymbols object which holds the index).

Queries the ProcessSymbols for matches but currently discards them and
returns not found. The selection of matches from the list will be done
in a separate pass to simplify the patch.

Marks more functions const that can be.

TEST=unit

Change-Id: I4ca635e711bbe9f1d68c19be4347d01a8b38e7c2