[pkg_url] fix pkg_url matching
It is not safe to perform partial matching in the way that was being done.
The result was that package_name was being set to foo#bar for a particular
case, breaking both component and binary resolution for shallowly placed
binaries.

The regex itself also now supports variants (so they are no longer rejected),
but they are not plumbed for external consumption.

Test: fuchsia_pkg_url_unittest
Change-Id: Iad6179bcd27146b71cd74e7a9381995de613fd2a