[refactor] component loader -> package loader
This change started with the goal of cleaning up some corner cases for
loading binaries. The code paths down in component loader, particularly
shared with the children were hairy and somewhat hard to follow, leading to
easy mistakes. In order to compact this, I ended up taking on a larger
refactor along the path.

Several things were learned, leading to the ultimate result:
 - component loader is a package loader, the naming was confusing.
 - the loader API is loading URLs that may or may not have resources.
 - the updating loader did not need a virtual below it, it just needed to
   override.
 - the special handling for file:// urls needed to be raised and isolated.

Test: fuchsia-pkg:// urls work, file:// urls work, /bin/rolldice works.
Change-Id: I648c4b5b7b02f511cc86e841eedfe856a507e735