[fbl->std] fbl::move -> std::move
fbl::move is completely gone *outside the kernel*.
Nothing has changed in the kernel yet.

fbl::move is identical to std::move but was not recognized specially by
the compiler.  Using std::move triggered diagnostic logic in the
compiler so many inappropriate uses were identified and removed.
(Roughly speaking, move should never be used on return values, neither
by a callee in return statement, nor by a caller of a function or
constructor using that return value in assignment or argument passing.)

Bug: ZX-1571 #comment Zircon userland fbl::move -> std::move
Test: CQ
Change-Id: I83f25a77aec74f43a6ac57218ab997bb7c2751df