[fbl->std] fbl::move -> std::move
fbl::move is identical to std::move but was not recognized specially by
the compiler.  Using std::move triggered diagnostic logic in the
compiler so many inappropriate uses were identified and removed.
(Roughly speaking, move should never be used on return values, neither
by a callee in return statement, nor by a caller of a function or
constructor using that return value in assignment or argument passing.)

Bug: ZX-1571 #comment Garnet fbl::move -> std::move
Test: CQ
Change-Id: I6e533ce60b4ad2e7db90ff4ebce4e4ea26011068