[mediaplayer] implement CQ test commands
This CL updates the mediaplayer CQ test to make it easier to write
tests that reproduce complex user behavior. The test framework now
implements a queue of commands that are executed asynchronously.
Commands include waiting for the player to reach a particular
position in the content and waiting a specified duration.

TEST: this is CQ test code...still passes
Change-Id: I917a70e28825d1643daed4b12724052ce46c362e