[dart] directly run component tests as components
If a dart test has a corresponding `foo.cmx` with the same name as the test -
i.e. it is a valid component test, then it can be run directly without the
shell stub.

runtests will invoke run_test_component if the test name matches the name of
a cmx file in meta. For these dart tests, that is always the case, so moving
the shell script to the same name as the CMX has the intended effect.

Change-Id: I795c1f37f39c9f453f6649e534bc47d853e9b23e