[msd-arm-mali] Transition L2 status on protected mode switch.
We need to power down the L2 before resetting the GPU to switch out of
protected mode so that all the caches are flushed.

Also stop trying to support protected mode if cache coherency is
enabled, because that requires some more complex mode-switching logic.

Also move power and reset interrupt processing to the interrupt thread,
because the device thread could be blocked on them in reset. We could
make reset asynchronous, but that would require that many other
operations could handle running during a reset.

Test: astro,vim2:go/magma-tps#L0
MA-485

Change-Id: I9813532196d658eeda2ab62778e61e458acdae20