[ethernet] Remove listnode from ethmac protocol
list_node_t should be private to each driver, so that it doesn't
accidently get double used in multiple layers of the driver stack. This
is following the same design nand and block protocols currently use, and
the one the usb protocols are transitioning to. The EthmacNetbuf struct
defines the public portion of the buffer, and a each driver in the stack
may define a private section afterwards. They report their size to
higher level stack via the Query call. The highest driver in the stack
(ethernet), is responsible for allocating a buffer at least as large as
it's device's parent.

Tested: Paved pixelbook, vim2, and astro
Change-Id: Ie066adbdf112e58d10ad114ce8fbdd78edecb513