[syscall][context-switch] Correct handling of debug state for arm64.
Implements the context switch for the debug state in arm64. This works
similarly to what x86 does, which simply copies over the state that is
validated beforehand.

Before the read_state syscall was getting the state of the hardware,
which is wrong because it's getting the state of the current thread, not
the one of the handle given. Now it returns the state that is saved with
the associated thread.

This also makes some minor improvements on the x86 for consistency
improvements. The tests cannot be activated because there are still data
races with the register state when suspension occurs.

TEST=unit tests passes.
ZX-2508  # Correctly handle debug state. FP state is still pending.
ZX-3025  # Done.

Change-Id: I52f9c5ed71020bb0ced4228a2a55595be9288645