[ktrace] Only record syscalls that have a start and end
Syscalls that either start and do not end or end and do not start
creates a confusing state for the trace processors.

Also as we match the start and end of syscall, we can to use a
more concise format when transformed into JSON if we chose to.

TEST: tested manually by generating trace with and without the
patch and verifying that only the edge syscall events are removed.

BUG:PT-64

Change-Id: I7b4c9128d7a261de94f5aee1d18ad48cf609887b