[mediaplayer] separate interactive tests from cq tests
This CL creates a separate binary and package for the manual test
app for mediaplayer vs the cq tests. The cq test is still called
mediaplayer_tests, but that binary only does hermetic tests for
cq, regression testing, etc. The new binary, mediaplayer_test_util,
is the command-line test that uses the ambient mediaplayer service
and shows simple UI.

This was done, because the hermetic test is now difficult to invoke
from the comment line without using run_test_component.

TEST: CQ test still passes, ran manual tests
Change-Id: Iecde3364f519f52c6ac2d4e27585d774b0e61324