[syscalls] Remove syzkaller references from garnet
syzkaller is no longer included in the main build, and is only built
elsewhere where it uses a copy that's not part of the main Fuchsia tree.

Remove this dependency in garnet for now to avoid confusion (and because
the current fork has uses of zx_task_resume() which we're in the process
of nuking).

ZX-2720 #comment

Test: CQ
Change-Id: I5fb1473891046c9e2d6beec8a951dd129aa24446