[zircon][usb][xhci] Avoid using req->node in lists.
Currently, there is only 1 list_node_t pointer in usb_request_t and
differnt drivers in the usb stack that track a usb request enqueue it
using this node. This will lead to bugs where the request is simultaneously
included in multiple lists changing the next and previous pointers of the node.
This changeset makes use of the 'node' from respective private contexts.

Test: Pixelbook- fx serve, iochk
Vim2- fx serve, iochk.
Astro- netboot, lsusb.

Change-Id: I5c44a2c274d900def4517a9a49a9f758e1583e69