[set] Guess product and board config from short names
This guesses the product and board configurations based on short names.
The heuristic it uses is that it matches the first entry of the form
$petal/{products,boards}/$shortname.gni starting at the current guessed
petal and going down through the petals in order until it finds a
matching file, or printing an error and exiting it if can't. This is
hopefully more ergonomic for developers than knowing exactly which
repository defines which product and board set.

Test: 'fx set x64 --product base --board frank' with various checkout
types.

Change-Id: I89c747b952bb89ca27d46c38c934f56d21dc5bd0