[bt-gap] Send stop_discovery FIDL message synchronously
The async declaration of HostDevice::stop_discovery() caused it to
never send a FIDL message to the bt-host device because stop_discovery()
is called from a drop() implementation that never polled the returned
future.

Fixed this by changing stop_discovery() to not be async so that its body
executes synchronously.

NET-1894 #done
Test: 1. bt-fake-hci add
      2. Run bt-cli -> start-discovery. Check logs for on-going inquiry
         messages.
      3. Run bt-cli -> stop-discovery. Inquiry procedures and on-going
         LE scans should stop.

Change-Id: I9f1431e44b4f56934d13efe3c3fe9eb16ed23bf1