[bt][bredr] Fix callback delay restarting discover
When the last discovery session was removed, the host continues to
perform inquiry until it completes

When this happens and a new session is added, it was not detected
that inquiry was active, so the request would be placed into a queue
that was completed after the inquiry would complete.  This meant clients
would wait up to 60 seconds to receive a callback even though inquiry
was occuring.

Also, lower the default inquiry duration so we do not continue after a
session is dropped for so long.

NET-1931 #done
Test: bt-host-unittests
 - GAP_BrEdrDiscoveryManagerTest.MultipleRequests
 - GAP_BrEdrDiscoveryManagerTest.RequestDiscoveryWhileStop

Change-Id: I9374bb2eacebbd53480abacb8f05b9eb4747da16