[debugger] Parse identifiers with "::"
Adds a "::" operator to the parser.

Adds an "Identifier" object to hold names of things that the parser
generates.

Currently the code just uses the full name out of the identifier so
there is almost no noticable change to users. in the future we'll be
able to traverse the scopes to resolve symbols in the index.

TEST=unit

Change-Id: I7de4e7f65e34048e4ccd5563410bc28dbf9c8939