[ethernet] Define MAC_HDR_MAX_SIZE in FIDL
The purpose of this change is to allow wlan drivers to reserve a small
amount of headroom (64 bytes) in every tx buffer so that it can prepend
its own headers, such as DataFrameHeader and LlcHeader.
This could eliminate the need to copy the data in the ethernet frame.

64 is the closest power of 2 that is larger than 44, which is
DataFrameHeader::max_size (36) + LlcHeader::max_size(8)

Test: Build and smoke test with wlan and ethernet. Value not used yet.

NET-1911 #comment

Change-Id: Ic88c8963d7ce7d8be15a4e26736df8cd116fcce8