[syscalls] Remove quiet unbinding, switch utest to zx_task_resume_from_exception()
Quiet unbinding doesn't currently work with
zx_task_resume_from_exception() (when exception port is checked) and it
is somewhat involved to make it work properly.

As it's not currently being used, remove
ZX_EXCEPTION_PORT_UNBIND_QUIETLY and related plumbing for now, to be
readded in a modified form in the future if necessary.

At the same time, switch utest/exception to stop using zx_task_resume()
in favour of zx_task_resume_from_exception(), as zx_task_resume() is
deprecated.

ZX-988 #comment
ZX-2720 #comment

Test: CQ, /boot/test/sys/exception-test, unbind_rebind_while_stopped_test removed.
Change-Id: If7f4706f7599c417cdc9447011441e42e892f0c2