[ledger] Initialize StartupContext to nullptr, require IOAsync
The pointer to the StartupContext was not initialized in the environment
builder, which made a debug assertion pass under normal conditions, but
fail with asan. This change sets it explicitly to nullptr by default,
and make the test set a valid environment.
This change also add an assertion to check the io_dispatcher is present,
changes the tests and update the comments to indicate what is required
when building an environment.

TEST=now passes with asan

Change-Id: I48dccdec0891f8333084eb81905f2a15cdcc34d4