[debugger] Parse identifiers with templates.
Adds a template list to the Identifier component (previously this was a
single opqaue string). The individual template parameters are still
opaque strings and are expected to remain that way.

Adds a template type extractor that can identify the range of a template
parameter and perform limited canonicalization. This will not support
understanding the full type declarations.

Updates the parser to extract the template parameters using the skipper
to produce the new identifier types. These identifiers are not hooked up
to anything so the main user-visible result will be different error
messages if you type an invalid template.

TEST=unit

Change-Id: I5f9f7079bd643809e0a3a2a7dd4503cac9c23780