[virtio][block] Make operations non-blocking and support flush
Note that a virtio flush request refers only to flushing volatile writes to
persistent storage wheres a flush operation in the block protocol is
also a barrier. Since the virtio specification allows devices to service
requests in any order, we must wait for pending transactions to complete
before and after making a flush request.

ZX-2724 #done

TEST=guest_integration_tests + fx run -V

Change-Id: I201688f8a468aca02d13c06948e25c52772ef4ce