[arm64][hypervisor] Use phys-ints for the vtimer
Use physical interrupts for the virtual timer, and correct the handling
of physical interrupts in the hypervisor.

We now mark physical interrupts as active on the physical distributor,
and only set HCR_EL2.VI if there is a pending virtual interrupt. In
addition, we use ELRSR more accurately to track the state of list
registers so we end up with fewer spurious interrupts.

We also unmask the virtual timer interrupt and the virtual maintenance
interrupt on each CPU, so that we can correctly respond to both by
exiting the guest.

There is still some work to be done to eliminate additional spurious
interrupts in Zircon guests.

MAC-180

Test: Ran "zircon_guest", "linux_guest", and "guest_integration_tests"
      on a VIM2.
Change-Id: I9b3eabffa28472ec5b285fe8812d0d61a1c2c59c