[fidl] Fix formatting for ordinal removal.
The AST has an invariant that the previous_end field of each node points to
the end of the previous AST node.  The previous CL deleted the AST node without
correctly updating the previous_end field of the subsequent node.  This had the result
of adding extra whitespace to the formatted code.

Test: Added a test that modifies the AST and formats it, and then reparses and formats the result to make sure they are the same.
Change-Id: Iabaff5e7a83ca69126f0b0edd66943a497c38b02