[debugger] SW Breakpoint Integration Test
This test required quite a bit of infrastructure.

The message loop changed the event handling a bit. The previous approach
check for C++ tasks and then watch handle events. But it didn't recheck
that the event handling posted more tasks, which is done through this
test.

In particular it needed to create a test .so and binary that will be
added to the debug agent package.

The debug agent package needed to be extended in order to be able to
access those files.

The test consist in looking into the .so symbol offset to be able to
know the offset within the binary that will be loaded as a process by
the debug agent. Then we can set a breakpoint and verify it was hit.

TEST=integration!

Change-Id: I8ae48344b0f4dfebfbb903fffc356384a15a7a32