[netstack_test] Filter Integration Test
This patchset is the first integration test for netstack's packet
filtering engine.

The end goal of this integration test it to exercise netstack's packet
filtering rules, eventually validating them by actually running traffic
against a given rule.

This first change adds a C++ launcher that starts a test instance of
netstack with a single dummy ethernet interface, and assigns a static
IP address of 192.168.250.1. It then launches a Rust test program that
actually modifies the packet filter rules. In an attempt to keep code
reviews small and focused, the bulk of the Rust parts of this patchset
will be added in a later commit. For the moment, the C++ launcher just
passes the static IP address that was added to the dummy ethertap
interface, and checks that the exit code is zero.

Test: Ran the integration test manually, and it passes.

Change-Id: Iad988fc7af78cf3b601596b1c8f595d0a4a277a0