[prebuilt_package] Fix two build dependency issues with prebuilt_package
This adds two missing build edges that may lead to inconsistent builds
when using prebuilt_packages:

1.) The primary build rule for a prebuilt_package produces the
$pkg.meta/meta.far.merkle file, but wasn't expressing this as an
output. This meant other rules that wanted to read this file did not
know that it was an output of this step. This patch declares this file
as an explicit output of the primary build rule, similar to how
package() does.

2.) The pkgsvr_index.rsp rule emitted references to files generated by
the primary build rule, but didn't declare a dependency on the primary
build rule. This meant rules that depended on the pkgsvr_index.rsp
rule, such as //build/images:pkgsvr_index, would potentially read files
generated by the primary build rule before or in parallel with the
primary build rule leading to inconsistent indices. This adds an
explicit dependency from the pkgsvr_index.rsp rule to the primary build
rule.

Test: Carefully compare 'gn desc' and 'fx build -v' outputs.
Change-Id: Icc439261f8abdd493d7b851679ac703c9f1aec91