[sysmem] Sysmem as a Zircon driver.
Because sysmem has a Zircon driver client (display driver) and a Zircon
process client (virtcon), sysmem needs to be a Zircon driver.  This CL
creates the sysmem driver and gets it loaded for all boards.

Future CLs will broker Zircon sysmem service requests to the driver, and
will broker garnet+ sysmem service requests to the driver (via the
Zircon service probably).

Currently the only way to contact the sysmem driver is to talk to
/dev/class/sysmem/000 directly.  See previous paragraph for how that
will change soon.

Some of the interfaces are being updated.  For now, the old sysmem
remains in place and still implements the old interfaces, but we'll move
those clients over soon.

For now, sysmem runs in the platform bus driver devhost (or ACPI
devhost) to permit sysmem to be contacted (soon) from devhost(s) that
are children of the platform bus driver devhost (the only way to do
driver-to-driver coms so far).  In particular, the display driver will
need to contact the sysmem driver.

Tested: sysmem_tests
Change-Id: I886ad89e6f166a136aa75f3f1128f71851928f0c