[wayland][scanner] Support protocol dependencies.
Allow protocols to depend on other protocols. Since this information is
not declared inside the protocol file, these must be declared in the
BUILD.gn. Implement this by simply brining all the symbols from the
dependent crate into scope, which is the same scoping rules applied by
the scanner.

MAC-183 #done

Test: linux-dmabuf-unstable-v1.xml compiles with the scanner.
Change-Id: Ifbe1f75a65f6024efb32ae776e618286d7aae205